core/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "core"
version = "0.1.0"
description = "Shared core components for LoL Coach"
dependencies = [
    "pydantic",
    "pyyaml",
    "python-dotenv",
]

[tool.setuptools.packages.find]
where = ["src"]
